added new file README.md
This git repository is intended to publish a collection of my
single file tools on GitHub. GitHub displays the contents of
a file called README.md on the repository page, if it exists.
The contents of the file are interpreted as Markdown formatted
text.

This README.md file shall list all the individual SFTs with a
short description and, if applicable, a link to the web page
for the tool.