rfc-reader: add two spaces to the left margin
This change improves readability for RFCs that do not use spaces on
the left margin, e.g. RFC 1034. The effect is similar to the two extra
columns on the right margin. As a side effect, this restores symmetry
between the left and right sides of the header and footer lines.