add initial version of renumber script
The script "renumber" is intended to create contiguous numbering of a
set of files that are numbered discontiguously. All files of the set
need have a common structure that can be described as

    <prefix><number><postfix>

with the number a positive integer consisting of a constant number of
digits.

The original numbering can start anywhere, the result will be numbered
starting at 1, retaining the original order of files and the number of
digits used for numbering.