rfc-reader: add empty page for "most" pager only
For pagers like "most" that add¹ an additional empty line after the
end of the file (EOF), fill the next page, but omit the line that
will be added by the pager.

If the original file lacks a Line Feed ('\f') at the end, i.e., the
last page is not filled by expanding the Line Feed, fill the last
page, but omit the lines added by the pager, if applicable.

Additionally, remove left over debug output and bump version number.

Footnotes:
¹ A text file conforming to POSIX ends with a Newline ('\n') character
  without actually starting a new line, but taking the file contents
  literally, another line that does not end with a Newline has been
  added by the Newline before End-of-File.