rfc-reader: remove a leading Unicode BOM
Text format RFCs containing UTF-8 characters are supposed to begin
with the UTF-8 encoded Unicode Byte Order Mark (BOM).  This use of
the BOM as format indicator as opposed to its original meaning as
byte order indicator for two-byte Unicode encodings is foreign to
GNU and Unix-like systems, where it is prone to cause problems.

Thus remove the UTF-8 encoded BOM at the beginning of the file before
displaying it in the pager.

The implementation in rfc-reader is non-portable and requires a
non-POSIX-crippled shell, e.g., Bash.  It requires use of a UTF-8
locale as well, but this is more or less required for correct
display of UTF-8 characters anyway.