new shell script archive_url
The shell script `archive_url` is intended to ensure that the given
URLs are archived by the Wayback Machine.

I want to use this to add every single issue of The Daily Dose of IETF
to the Wayback Machine before the server it ran on is turned off.
There are 3811 issues of The Daily Dose of IETF, but only a random
subset was archived already, and none of the individual issue URLs
were archived before this weekend.