pmtud: use IPv4 by default, allow use of IPv6
Path MTU discovery for IPv6 using "ping" does not work with at
least some "ping" implementations on GNU/Linux.  Packets bigger
than the local interface MTU are fragmented before sending.

The pmtud script attempts to determine the outgoing interface
MTU and thus should often avoid that issue.

The outgoing interface MTU can only be determined when the target
is given as an IPv4 or IPv6 address, not when given as a hostname.