create_ipv6_ULA_global_prefix_rfc4193.sh: Ubuntu 18.04 support
On Ubuntu 18.04 GNU/Linux, there is neither ntpq installed by default,
nor is any EUI-64 based IPv6 address created by default.

If ntpq is not available, attempt to use date (this works with GNU date)
to determine the current time.

If no EUI-64 based IPv6 address is available, look for a MAC address,
with a fallback based on uname, to determine a "suitably unique identifier,
local to the node".

Add a couple of "defensive" double-quotes as well.