new shell/awk script top-level-section
The shell/awk script `top-level-section` is a variation on the
*section* theme.  While the three `section` implementations in
SFT are patterned after the *Cisco IOS* output modifier `section`,
`top-level-section` is patterned after the *Arista EOS* output
modifier `section`.

The current *top-level* is defined as the line with the minimum
amount of indentation encountered so far.  A section always
starts at the current top-level and comprises all directly
following lines with deeper indentation.  When a given line
matches the provided regular expression, the current top-level
section is printed completely.

This `top-level-section` program does not implement the special
treatment of *separator lines* in the configuration of Arista
EOS switches by the Arista EOS `section` output modifier.