section: change last "ind" to "ind_len"
This was an error, because "ind" was no longer used, because it
had been replaced with "ind_len."  But resetting "ind" or "ind_len"
outside a section is not really necessary, because it is set upon
entering a section and used only when already inside one.  Thus
this was not a user-visible bug, AFAICT.  Fix it anyway.