new shell/awk script enclosing-section
The shell/awk script `enclosing-section` is a variation on the
*section* theme.  While the three `section` implementations in
SFT are patterned after the *Cisco IOS* output modifier `section`,
`enclosing-section` is patterned after the *Huawei VRP* output
modifier `section include`.

The *enclosing* section is defined to start at the first line
preceding the matching line with less indentation, or the matching
line if no such line exists.  The section is defined as the start
line and all directly following lines with more indentation.

This script has only been lightly tested, please beware of bugs.