net2ips: avoid error for /31 networks
Before, net2ips would try to evalute the program "n" when given
a /31 CIDR prefix and using my SFT ipcalc.  It no longer does
this.

Printing the IP addresses of a /31 network still does not work,
because net2ips unconditionally removes the first and last IP
address, which is wrong for /31 networks.