vrp_output_to_yaml.awk: support firewall clusters
Huawei VRP firewalls in cluster mode prefix the prompt to
show that this is a firewall cluster and whether it is the
active or standby node.  Accept this prefixed prompt as well
as the router or switch prompt without prefix.

This is based on a report with patch from Sebastian Deußer.