vrp_output_to_yaml.awk: define prompt regex once
Instead of having four copies of the definition of the
regular expression matching the VRP device prompt, use
a variable to hold the regular expression.

The variable contents are defined once at the start of
the script (in the BEGIN block).  This variable is used
every time the device prompt needs to be matched.