reachability_monitor: adjust FreeBSD ping options
Testing on macOS showed that "ping -c1" would wait for the
sum of both "-i" and "-W", i.e., "ping -c1 -W1000" would
wait for 2s instead of 1.  Thus set the minimum -i value
allowed for non-root users (0.1s) and use a wait time of
900ms for a total of 1s.

This may work around a bug in the ping version of macOS,
in which case I'd probably split the "freebsd" ping option
set name into "macos" and "freebasd" with different options.
But I would need a test report from FreeBSD or access to
a FreeBSD system to test this myself first.