ext_ip: use wget or curl for icanhazip
While all previously used IP address services returned
an answer using HTML, icanhazip returns just text without
markup.  Thus both wget and curl suffice for icanhazip.

Prefer wget and curl for requests to icanhazip, and skip
any IP address service that needs a tool that is not
available.

This script changes from POSIX compatible sh to Bash in
order to use arrays, because using arrays allows quoting
of a variable that is expanded into mutliple words.
This is used to select the appropriate options to use
for the three tools lynx, wget, and curl.  This might be
a problem for BusyBox based systems without Bash.

Setting the HAVE_LYNX and HAVE_WGET variables to the
selected tool is intended to possibly help debugging.