reachability_monitor: prevent normal use
The reachability_monitor script does not always clean
up after itself, i.e., it may leave host monitors
running after the main script has exited.  But just
adding a call to stop_monitors() sometimes results
in the error that one of the PIDs to kill does not
exist (this PID may even be used by a different process).
Thus this is not yet ready for general use.

If you know what you are doing and I can trust you,
you can still use the script, but you have to ensure
that everything is cleaned up afterwards.