ensure_bom_crlf: new option -r
Using the option "-r" reverses the operation of ensure_bom_crlf.
Instead of ensuring that the file starts with a so called "BOM"
and uses <CR><LF> to separate lines, it ensures that the file
does not start with a so called "BOM" and that it uses <LF> to
separate lines.