rfc-reader: avoid trailing empty page
Some documents, e.g., RFC 868 and BCP 41, end with a trailing
empty line after the last Form Feed line.  Before, this was
expanded to a trailing empty page.  Now, this trailing empty
line is not printed, avoiding a trailing empty page.

As a side effect of the implementation, all blank lines are
replaced with an empty line (i.e., just a newline) on output.