ext_ip: Wget and curl work with DynDNS, too
The sed script used to extract the IP address from the DynDNS
web page rendered with lynx works for the raw HTML document
returned via Wget or curl as well.  Thus use Wget or curl if
available for DynDNS, with a fallback to lynx otherwise.
(This is the same situation as with "icanhazip".)