ext_ip: remove lynx-only remnants
ShellCheck complained about the unused variables LYNX and
HAVE_LYNX.  They are remnants from when I used lynx to parse
returned HTML instead of just extracting an IPv4 or IPv6
address.  This was useful for some "what is my IP?" services,
but not needed for the current set.  I also know of additional
such services that return simple answers and do not require
rendering the HTML to text.  Thus, remove the unused variables,
and add a fallback to lynx to the code that prefers Wget and
can already fall back to curl.