rfc-reader: stricter I-D version interpretation
According to the "Internet-Draft Author Resoucres" wiki[1],
"Internet-Drafts are versioned with two digits, separated
from the name with a hyphen."  Require exactly this when
determining whether the provided argument includes the exact
version number or not.

Before, "draft-kumkova-v6ops-nat64-wkp-1918" was interpreted
as specifying version "-1918".  Now, this is interpreted as
having no version number, resulting in rfc-reader attempting
to find the latest version of the draft (currently "-00").

[1] <https://authors.ietf.org/en/naming-your-internet-draft>
    (checked 2025-11-03)